agro_api/requirements.txt
Django>=4.0
psycopg2-binary
requests
gunicorn == 22.0.0